Tighten investigation lifecycle and separate mutation outcomes
